fix: stabilize nominal union tie breaks

+++ crates/sifr_codegen/src/lib_codegen_tests/structural_impl_demand_codegen_tests.rs
@@ -189,7 +189,7 @@ fn project_root_record_keeps_qualified_structural_identity() {
         &[],
         &[NominalField {
             name: "value",
-            identity: primitive("i64"),
+            identity: primitive("int"),
             required: true,
             default_identity: None,
         }],
